+++ include/writeNode.h
@@ -18,6 +18,6 @@ WriteNode* NewWriteNode(int padding, int fontSize, int x, int y);
 void SetWriteNodeSize(WriteNode* node, int padding, int fontSize);
 void SetWriteNodePosition(WriteNode* node, int x, int y);
 void DrawWriteNode(WriteNode* node);
-int EvaluateWriteNode(WriteNode* node, Dictionary* dict); // TODO: functiile astea ar tb sa returneze rezultatul direct, iar alta functie sa verifice corectitudinea fiecarei expresii din fiecare nod inainte de a executa
+float EvaluateWriteNode(WriteNode* node, Dictionary* dict); // TODO: functiile astea ar tb sa returneze rezultatul direct, iar alta functie sa verifice corectitudinea fiecarei expresii din fiecare nod inainte de a executa
 void ResizeWriteNode(WriteNode* node);
 void GetInputWriteNode(WriteNode* node);
